Refactoring game rules and options into a class module
