+++ views/index.erb
@@ -2,7 +2,7 @@
 
 <h2>Rules</h2>
 <ul>
-  <% (Game.rules).each do |rule| %>
+  <% (Game::RulesOptions::rules).each do |rule| %>
     <li><%= rule %></li>
   <% end %>
 </ul>
